MobileApp/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
boto3==1.34.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
Pillow==10.2.0
langdetect==1.0.9
mangum==0.17.0
pytest==7.4.4
pytest-asyncio==0.23.3
